Rename demo group to MotorTransactionView
