+++ Demo/Demo/Fullscreen/MotorTransactionView/DemoViewModels/BuyerProcess/WithInsuranceConfirmationDemoViewModel.swift
@@ -4,9 +4,9 @@
 
 import Foundation.NSString
 
-extension TransactionDemoViewDefaultData {
-    static var ContractNotCreatedDemoViewModel = MotorTransactionModel(
-        title: "Salgsprosess",
+extension MotorTransactionDefaultData {
+    static var WithInsuranceConfirmationDemoViewModel = MotorTransactionModel(
+        title: "Kjøpsprosess",
 
         header: MotorTransactionHeaderModel(
             adId: "171529672",
@@ -17,9 +17,8 @@ extension TransactionDemoViewDefaultData {
         steps: [
             MotorTransactionStepModel(
                 state: .completed,
-                style: .default,
                 main: MotorTransactionStepContentModel(
-                    title: "Annonsen er lagt ut",
+                    title: "Du har funnet en bil",
                     primaryButton: MotorTransactionButtonModel(
                         text: "Se annonsen",
                         style: "FLAT",
@@ -27,31 +26,45 @@ extension TransactionDemoViewDefaultData {
                         fallbackUrl: "/171529672"))),
 
             MotorTransactionStepModel(
-                state: .active,
+                state: .completed,
                 style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Kontrakt",
-                    body: NSAttributedString(string: "Når du har funnet en kjøper er det neste steget å skrive en kontrakt."),
+                    body: NSAttributedString(string: "Begge har signert kontrakten"),
                     primaryButton: MotorTransactionButtonModel(
-                        text: "Opprett digital kontrakt",
-                        style: "CALL_TO_ACTION",
+                        text: "Gå til kontrakt",
+                        style: "FLAT",
                         fallbackUrl: "https://www.google.com/search?q=contract+signed"))),
 
             MotorTransactionStepModel(
-                state: .notStarted,
+                state: .completed,
+                style: .default,
+                main: MotorTransactionStepContentModel(
+                    title: "Forsikring",
+                    body: NSAttributedString(string: "Du har oppgitt at du har forsikret denne bilen.")),
+                detail: MotorTransactionStepContentModel(
+                    body: NSAttributedString(string: "Om du vil bytte forsikring, kan du finne en ny som passer deg på våre sider."),
+                    primaryButton: MotorTransactionButtonModel(
+                        text: "Se bilforsikringer",
+                        style: "FLAT",
+                        action: "URL",
+                        url: "https://www.google.com/search?q=contract+signed"))),
+
+            MotorTransactionStepModel(
+                state: .active,
                 main: MotorTransactionStepContentModel(
                     title: "Betaling",
-                    body: NSAttributedString(string: "Dere kan betale trygt gjennom FINN ved å velge det i kontrakten."))),
+                    body: NSAttributedString(string: "Etter at du har forsikret bilen kan du starte betalingen"))),
 
             MotorTransactionStepModel(
                 state: .notStarted,
+                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Overlevering",
                     body: NSAttributedString(string: "<p>Velger dere å betale gjennom FINN, må overleveringen skje innen 7 dager etter kjøper har betalt.</p><p>Registrering av eierskiftet bør gjøres når dere møtes for overlevering.</p>"))),
 
             MotorTransactionStepModel(
                 state: .notStarted,
-                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Gratulerer med salget!",
                     body: NSAttributedString(string: "Du kan finne igjen bilen i Mine kjøretøy under «Eide før»."))),
