+++ Demo/Demo/Fullscreen/MotorTransactionView/DemoViewModels/SalesProcess/AdExpiredDemoViewModel.swift
@@ -1,11 +1,11 @@
 //
-//  Created by Saleh-Jan, Robin on 28/02/2020.
+//  Copyright © 2020 FINN AS. All rights reserved.
 //
 
 import Foundation.NSString
 
-extension TransactionDemoViewDefaultData {
-    static var SellerOnlySignedDemoViewModel = MotorTransactionModel(
+extension MotorTransactionDefaultData {
+    static var AdExpiredDemoViewModel = MotorTransactionModel(
         title: "Salgsprosess",
 
         header: MotorTransactionHeaderModel(
@@ -14,33 +14,35 @@ extension TransactionDemoViewDefaultData {
             registrationNumber: "CF40150",
             imagePath: "2020/2/vertical-0/26/2/171/529/672_525135443.jpg"),
 
-        alert: MotorTransactionAlertModel(
-            title: "Du har opprettet flere kontrakter for denne bilen",
-            message: "En avtale er bindene når begge har signert. Prosessen under viser derfor prosessen for den første kontrakten begge signerte.",
-            imageIdentifier: "MULTIPLE_CONTRACTS"
-        ),
-
         steps: [
             MotorTransactionStepModel(
-                state: .completed,
-                style: .default,
+                state: .active,
+                style: .warning,
                 main: MotorTransactionStepContentModel(
-                    title: "Annonsen er lagt ut",
+                    title: "Annonsen er utløpt",
+                    body: NSAttributedString(string: "Legg ut annonsen på nytt, sånn at kjøpere kan ta kontakt med deg."),
                     primaryButton: MotorTransactionButtonModel(
-                        text: "Se annonsen",
-                        style: "FLAT",
-                        action: "SEE_AD",
-                        fallbackUrl: "/171529672"))),
+                        text: "Legg ut på nytt",
+                        style: "CALL_TO_ACTION",
+                        action: "REPUBLISH_AD",
+                        fallbackUrl: "/user/ads/admin.html?finnkode=171529672")),
+                detail: MotorTransactionStepContentModel(
+                    title: "Ønsker du hjelp med salget?",
+                    body: NSAttributedString(string: "Nettbil hjelper deg kostnadsfritt med å selge bilen til forhandlere gjennom en budrunde. De trenger bare levere deg bilen, så ordner Nettbil resten."),
+                    primaryButton: MotorTransactionButtonModel(
+                        text: "Mer om Nettbil",
+                        style: "DEFAULT",
+                        fallbackUrl: "https://www.finn.no/nettbil/velkommen"))),
 
             MotorTransactionStepModel(
-                state: .active,
+                state: .notStarted,
                 style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Kontrakt",
-                    body: NSAttributedString(string: "Venter på at kjøper skal signere."),
+                    body: NSAttributedString(string: "Når du har funnet en kjøper er det neste steget å skrive en kontrakt."),
                     primaryButton: MotorTransactionButtonModel(
-                        text: "Gå til kontrakt",
-                        style: "DEFAULT",
+                        text: "Opprett digital kontrakt",
+                        style: "FLAT",
                         fallbackUrl: "https://www.google.com/search?q=contract+signed"))),
 
             MotorTransactionStepModel(
@@ -51,13 +53,13 @@ extension TransactionDemoViewDefaultData {
 
             MotorTransactionStepModel(
                 state: .notStarted,
-                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Overlevering",
                     body: NSAttributedString(string: "<p>Velger dere å betale gjennom FINN, må overleveringen skje innen 7 dager etter kjøper har betalt.</p><p>Registrering av eierskiftet bør gjøres når dere møtes for overlevering.</p>"))),
 
             MotorTransactionStepModel(
                 state: .notStarted,
+                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Gratulerer med salget!",
                     body: NSAttributedString(string: "Du kan finne igjen bilen i Mine kjøretøy under «Eide før»."))),
