+++ Demo/Demo/Fullscreen/MotorTransactionView/DemoViewModels/SalesProcess/BuyerConfirmedHandoverDemoViewModel.swift
@@ -4,8 +4,8 @@
 
 import Foundation.NSString
 
-extension TransactionDemoViewDefaultData {
-    static var PaymentCompletedDemoViewModel = MotorTransactionModel(
+extension MotorTransactionDefaultData {
+    static var BuyerConfirmedHandoverDemoViewModel = MotorTransactionModel(
         title: "Salgsprosess",
 
         header: MotorTransactionHeaderModel(
@@ -23,6 +23,7 @@ extension TransactionDemoViewDefaultData {
         steps: [
             MotorTransactionStepModel(
                 state: .completed,
+                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Annonsen er lagt ut",
                     primaryButton: MotorTransactionButtonModel(
@@ -33,17 +34,17 @@ extension TransactionDemoViewDefaultData {
 
             MotorTransactionStepModel(
                 state: .completed,
+                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Kontrakt",
                     body: NSAttributedString(string: "Begge har signert kontrakten."),
                     primaryButton: MotorTransactionButtonModel(
-                    text: "Gå til kontrakt",
-                    style: "FLAT",
-                    url: "https://www.google.com/search?q=contract+signed"))),
+                        text: "Gå til kontrakt",
+                        style: "FLAT",
+                        url: "https://www.google.com/search?q=contract+signed"))),
 
             MotorTransactionStepModel(
                 state: .completed,
-                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Betaling",
                     body: NSAttributedString(string: "Kjøper betalte 1. februar 2020.")),
@@ -55,7 +56,7 @@ extension TransactionDemoViewDefaultData {
                 style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Overlevering",
-                    body: NSAttributedString(string: "<p>Dere må bekrefte før:<br/><strong>8. februar 2020.</strong></p><ol><li>Ved oppmøte registrerer dere først eierskiftet digitalt hos Statens vegvesen.</li><li>Deretter må <strong>begge</strong> bekrefte at overleveringen har skjedd, og at pengene kan utbetales.</li></ol>"),
+                    body: NSAttributedString(string: "<p>Kjøper har bekreftet.</p><p>Dere må bekrefte før:<br/><strong>8. februar 2020.</strong></p><ol><li>Ved oppmøte registrerer dere først eierskiftet digitalt hos Statens vegvesen.</li><li>Deretter må <strong>begge</strong> bekrefte at overleveringen har skjedd, og at pengene kan utbetales.</li></ol>"),
                     nativeButton: MotorTransactionButtonModel(
                         text: "Registrer eierskifte",
                         style: "DEFAULT",
