+++ Demo/Demo/Fullscreen/MotorTransactionView/DemoViewModels/SalesProcess/BuyerInvitedDemoViewModel.swift
@@ -1,11 +1,11 @@
 //
-//  Created by Saleh-Jan, Robin on 28/02/2020.
+//  Copyright © 2020 FINN AS. All rights reserved.
 //
 
 import Foundation.NSString
 
-extension TransactionDemoViewDefaultData {
-    static var BothPartiesConfirmedHandoverDemoViewModel = MotorTransactionModel(
+extension MotorTransactionDefaultData {
+    static var BuyerInvitedDemoViewModel = MotorTransactionModel(
         title: "Salgsprosess",
 
         header: MotorTransactionHeaderModel(
@@ -23,7 +23,6 @@ extension TransactionDemoViewDefaultData {
         steps: [
             MotorTransactionStepModel(
                 state: .completed,
-                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Annonsen er lagt ut",
                     primaryButton: MotorTransactionButtonModel(
@@ -33,39 +32,33 @@ extension TransactionDemoViewDefaultData {
                         fallbackUrl: "/171529672"))),
 
             MotorTransactionStepModel(
-                state: .completed,
-                main: MotorTransactionStepContentModel(
-                    title: "Kontrakt",
-                    body: NSAttributedString(string: "Begge har signert kontrakten."),
-                    primaryButton: MotorTransactionButtonModel(
-                        text: "Gå til kontrakt",
-                        style: "FLAT",
-                        url: "https://www.google.com/search?q=contract+signed"))),
-
-            MotorTransactionStepModel(
-                state: .completed,
-                main: MotorTransactionStepContentModel(
-                    title: "Betaling",
-                    body: NSAttributedString(string: "Kjøper betalte 1. februar 2020."))),
-
-            MotorTransactionStepModel(
-                state: .completed,
-                main: MotorTransactionStepContentModel(
-                    title: "Overlevering",
-                    body: NSAttributedString(string: "Dere har bekreftet at overleveringen har skjedd."))),
-
-            MotorTransactionStepModel(
                 state: .active,
                 style: .default,
                 main: MotorTransactionStepContentModel(
+                    title: "Kontrakt",
+                    body: NSAttributedString(string: "Kjøper har blitt invitert."),
+                    primaryButton: MotorTransactionButtonModel(
+                        text: "Signer kontrakt",
+                        style: "CALL_TO_ACTION",
+                        fallbackUrl: "https://www.google.com/search?q=contract+signed"))),
+
+            MotorTransactionStepModel(
+                state: .notStarted,
+                main: MotorTransactionStepContentModel(
+                    title: "Betaling",
+                    body: NSAttributedString(string: "Dere kan betale trygt gjennom FINN ved å velge det i kontrakten."))),
+
+            MotorTransactionStepModel(
+                state: .notStarted,
+                style: .default,
+                main: MotorTransactionStepContentModel(
+                    title: "Overlevering",
+                    body: NSAttributedString(string: "<p>Velger dere å betale gjennom FINN, må overleveringen skje innen 7 dager etter kjøper har betalt.</p><p>Registrering av eierskiftet bør gjøres når dere møtes for overlevering.</p>"))),
+
+            MotorTransactionStepModel(
+                state: .notStarted,
+                main: MotorTransactionStepContentModel(
                     title: "Gratulerer med salget!",
-                    body: NSAttributedString(string: "Du kan finne igjen bilen i Mine kjøretøy under «<a href=\"/minekjoretoy\">Eide før</a>»."),
-                    nativeBody: NSAttributedString(string: "Du kan finne igjen bilen i Mine kjøretøy under Eide før.")),
-                detail: MotorTransactionStepContentModel(
-                    body: NSAttributedString(string: "Det kan ta noen dager før pengene dukker opp på kontoen din."),
-                    nativeButton: MotorTransactionButtonModel(
-                        text: "Gå til Mine kjøretøy",
-                        style: "FLAT",
-                        url: "/minekjoretoy"))),
+                    body: NSAttributedString(string: "Du kan finne igjen bilen i Mine kjøretøy under «Eide før»."))),
     ])
 }
