+++ Demo/Demo/Fullscreen/MotorTransactionView/DemoViewModels/SalesProcess/ContractCreatedDemoViewModel.swift
@@ -4,15 +4,9 @@
 
 import Foundation.NSString
 
-extension TransactionDemoViewDefaultData {
-    static var BuyerInvitedDemoViewModel = MotorTransactionModel(
+extension MotorTransactionDefaultData {
+    static var ContractCreatedDemoViewModel = MotorTransactionModel(
         title: "Salgsprosess",
-
-        header: MotorTransactionHeaderModel(
-            adId: "171529672",
-            title: "BMW i3",
-            registrationNumber: "CF40150",
-            imagePath: "2020/2/vertical-0/26/2/171/529/672_525135443.jpg"),
 
         alert: MotorTransactionAlertModel(
             title: "Du har opprettet flere kontrakter for denne bilen",
@@ -36,9 +30,9 @@ extension TransactionDemoViewDefaultData {
                 style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Kontrakt",
-                    body: NSAttributedString(string: "Kjøper har blitt invitert."),
+                    body: NSAttributedString(string: "Du har opprettet kontrakt."),
                     primaryButton: MotorTransactionButtonModel(
-                        text: "Signer kontrakt",
+                        text: "Inviter kjøper",
                         style: "CALL_TO_ACTION",
                         fallbackUrl: "https://www.google.com/search?q=contract+signed"))),
 
@@ -50,13 +44,13 @@ extension TransactionDemoViewDefaultData {
 
             MotorTransactionStepModel(
                 state: .notStarted,
-                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Overlevering",
                     body: NSAttributedString(string: "<p>Velger dere å betale gjennom FINN, må overleveringen skje innen 7 dager etter kjøper har betalt.</p><p>Registrering av eierskiftet bør gjøres når dere møtes for overlevering.</p>"))),
 
             MotorTransactionStepModel(
                 state: .notStarted,
+                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Gratulerer med salget!",
                     body: NSAttributedString(string: "Du kan finne igjen bilen i Mine kjøretøy under «Eide før»."))),
