+++ Demo/Demo/Fullscreen/MotorTransactionView/DemoViewModels/SalesProcess/PaymentCompletedDemoViewModel.swift
@@ -4,9 +4,9 @@
 
 import Foundation.NSString
 
-extension TransactionDemoViewDefaultData {
-    static var BuyerConfirmedHandoverDemoViewModel = MotorTransactionModel(
-        title: "Kjøpsprosess",
+extension MotorTransactionDefaultData {
+    static var PaymentCompletedDemoViewModel = MotorTransactionModel(
+        title: "Salgsprosess",
 
         header: MotorTransactionHeaderModel(
             adId: "171529672",
@@ -23,7 +23,6 @@ extension TransactionDemoViewDefaultData {
         steps: [
             MotorTransactionStepModel(
                 state: .completed,
-                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Annonsen er lagt ut",
                     primaryButton: MotorTransactionButtonModel(
@@ -34,17 +33,17 @@ extension TransactionDemoViewDefaultData {
 
             MotorTransactionStepModel(
                 state: .completed,
-                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Kontrakt",
                     body: NSAttributedString(string: "Begge har signert kontrakten."),
                     primaryButton: MotorTransactionButtonModel(
-                        text: "Gå til kontrakt",
-                        style: "FLAT",
-                        url: "https://www.google.com/search?q=contract+signed"))),
+                    text: "Gå til kontrakt",
+                    style: "FLAT",
+                    url: "https://www.google.com/search?q=contract+signed"))),
 
             MotorTransactionStepModel(
                 state: .completed,
+                style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Betaling",
                     body: NSAttributedString(string: "Kjøper betalte 1. februar 2020.")),
@@ -56,7 +55,7 @@ extension TransactionDemoViewDefaultData {
                 style: .default,
                 main: MotorTransactionStepContentModel(
                     title: "Overlevering",
-                    body: NSAttributedString(string: "<p>Kjøper har bekreftet.</p><p>Dere må bekrefte før:<br/><strong>8. februar 2020.</strong></p><ol><li>Ved oppmøte registrerer dere først eierskiftet digitalt hos Statens vegvesen.</li><li>Deretter må <strong>begge</strong> bekrefte at overleveringen har skjedd, og at pengene kan utbetales.</li></ol>"),
+                    body: NSAttributedString(string: "<p>Dere må bekrefte før:<br/><strong>8. februar 2020.</strong></p><ol><li>Ved oppmøte registrerer dere først eierskiftet digitalt hos Statens vegvesen.</li><li>Deretter må <strong>begge</strong> bekrefte at overleveringen har skjedd, og at pengene kan utbetales.</li></ol>"),
                     nativeButton: MotorTransactionButtonModel(
                         text: "Registrer eierskifte",
                         style: "DEFAULT",
