+++ Demo/Demo/Fullscreen/MotorTransactionView/DemoViewModels/SalesProcess/SellerConfirmedHandoverDemoViewModel.swift
@@ -4,7 +4,7 @@
 
 import Foundation.NSString
 
-extension TransactionDemoViewDefaultData {
+extension MotorTransactionDefaultData {
     static var SellerConfirmedHandoverDemoViewModel = MotorTransactionModel(
         title: "Salgsprosess",
